etapas 5 e 6
